+++ components/_utils/cls.ts
@@ -1,11 +1,5 @@
 type ClassValue =
-  | string
-  | number
-  | boolean
-  | undefined
-  | null
-  | Record<string, boolean | undefined | null>
-  | ClassValue[]
+  string | number | boolean | undefined | null | Record<string, boolean | undefined | null> | ClassValue[]
 
 export function cls(...args: ClassValue[]): string {
   const classes: string[] = []